Moved Product & OrderItem implementation to traits
