+++ src/Entity/OrderItemInterface.php
@@ -12,24 +12,9 @@ declare(strict_types=1);
 
 namespace Brille24\SyliusCustomerOptionsPlugin\Entity;
 
-use Sylius\Component\Core\Model\OrderItemInterface as BaseOrderItemInterface;
+use Brille24\SyliusCustomerOptionsPlugin\Traits\OrderItemCustomerOptionCapableTraitInterface;
+use Sylius\Component\Core\Model\OrderItemInterface as SyliusOrderItemInterface;
 
-interface OrderItemInterface extends BaseOrderItemInterface
+interface OrderItemInterface extends SyliusOrderItemInterface, OrderItemCustomerOptionCapableTraitInterface
 {
-    /**
-     * @param OrderItemOptionInterface[] $customerOptionConfiguration
-     */
-    public function setCustomerOptionConfiguration(array $customerOptionConfiguration): void;
-
-    /**
-     * @return OrderItemOptionInterface[]
-     */
-    public function getCustomerOptionConfiguration(): array;
-
-    /**
-     * Returns the compressed version of the customer option configuration as though it was sent through a web-request.
-     *
-     * @return array
-     */
-    public function getCustomerOptionConfigurationAsSimpleArray(): array;
 }
